+++ packages/expo-dev-menu/android/src/main/java/expo/modules/devmenu/interfaces/DevMenuManagerProviderInterface.kt
@@ -1,0 +1,9 @@
+package expo.modules.devmenu.interfaces
+
+@FunctionalInterface
+interface DevMenuManagerProviderInterface {
+  /**
+   * Provides access to the object implementing the [DevMenuManagerInterface] interface.
+   */
+  fun getDevMenuManager(): DevMenuManagerInterface
+}